+++ src/scorpio.h
@@ -1189,9 +1189,9 @@ typedef struct PROCESSOR {
     static uint32_t eval_hash_tab_mask;
     static int age;
 
-    void  reset_hash_tab(int id,uint32_t = 0);
-    void  reset_pawn_hash_tab(uint32_t = 0);
-    void  reset_eval_hash_tab(uint32_t = 0);
+    void  reset_hash_tab(int id,size_t = 0);
+    void  reset_pawn_hash_tab(size_t = 0);
+    void  reset_eval_hash_tab(size_t = 0);
     void  delete_hash_tables();
     static void  clear_hash_tables();
     static int hashfull();
